Add new tile assets, sprites, and palettes for 2D ISO
Added new tile assets and corresponding sprites (maisonX256, nouvelle_resolution, nouvelle_resolutionV2, sol_herbeux, touch_grass) and their meta files. Introduced new palettes (New Palette, New Palette 1) and updated PaletteSol.prefab to include new tiles and sprites. Also added a new UI script CanvaManger and made various updates to event system and manager scripts.

+++ Assets/Script/Relation/GodDataSO.cs
@@ -11,7 +11,7 @@ public class GodDataSO : ScriptableObject
     public bool unlocked = false;
     public int relation = 0;
     // Interaction tracking
-    public int lastInteractionDay = 0;
+    public float lastInteractionDay = 0;
     public int interactionsToday = 0;
 
     [Header("Dialogues (petits graphs)")]
